requirements.txt
numpy
tensorflow
Pillow